Add automatic regtest step for the holder commit step (first successful such ordinal swap)
In the ordinal swap, the commit step for the holder is not used. Instead there is a manual payment
that the holder makes to the swap address with the ordinal. This version of the automation implements a
commit step to carry out the creation and payment of the ordinal, using the dev harness. The goal is to
enable integration tests to be made. In a subsequent commit, the code to create the ordinal is to be
moved to a separate step so that the ordinal location may be automatically placed into regtest orders.
In this version, there are two ordinal locations created, the second of which is used in the dev harness
form in the next swap. This step, once moved, will be kept separate from the order creation so that more
sophisticated automatic regtest ordinal creation may be implemented in the future, say with an ord server
inscribing the ordinals. Also there is existing regtest mining code in the holder commit step, a copy of
which can be used at the end of the swap to confirm the final transaction. So there are two tasks to be
done yet in preparation for creating an integration test for the ordinal swap, and the code for these
has already been exercised.